+++ CHANGELOG.md
@@ -10,8 +10,9 @@
     - Added custom tags for used-defined module grouping
     - Added auto-tagging, including MetaModule
     - Added fuzzy search (similar to the default module browser)
-    - Added option to select "Favorite" handling  (Legacy (MB) / (Built-in) VCV Rack)
-    - Added option to highlight favorites
+    - Added an option to select "Favorite" handling  (Legacy (MB) / (Built-in) VCV Rack)
+    - Added an option to highlight favorites
+    - Added an option to set the fuzzy search threshold
     - Added "v2 mod" variation
     - Changed hotkey to toggle hidden modules to Shift+Space (because of Spotlight on Mac)
     - Fixed broken button of "Favorites" category